+++ TML.Patcher.Frontend/Program.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Reflection;
 using System.Threading;
 using Consolation.Common.Framework.OptionsSystem;
+using Microsoft.Win32;
 using TML.Patcher.Frontend.Common;
 using TML.Patcher.Frontend.Common.Options;
 
@@ -28,8 +29,15 @@ namespace TML.Patcher.Frontend
 
             Consolation.Consolation.Window = new Patcher(args);
 
+            PreLoadAssemblies();
+            Patcher.InitializeConsoleOptions();
+            Patcher.InitializeProgramOptions();
+
             if (Configuration.ShowIlSpyCmdInstallPrompt)
                 InstallILSpyCMD();
+
+            if (Configuration.ShowRegistryAdditionPrompt)
+                AddRegistryContext();
 
             ConfigurationFile.Save();
 
@@ -100,6 +108,46 @@ namespace TML.Patcher.Frontend
 
             process.Start();
             process.WaitForExit();
+        }
+
+        private static void AddRegistryContext()
+        {
+            if (!OperatingSystem.IsWindows())
+                return;
+
+            Patcher window = Consolation.Consolation.GetWindow<Patcher>();
+
+            window.WriteLine("Do you want to add TML.Patcher.Frontend to your file context menu? Please ensure that this program is located in a location it will not move from.");
+            window.WriteLine("<y/n> (or 'p' to skip for now and preserve the prompt for later)");
+
+            ConsoleKeyInfo pressedKey = Console.ReadKey();
+            window.WriteLine();
+
+            if (pressedKey.Key == ConsoleKey.P)
+                return;
+
+            if (pressedKey.Key != ConsoleKey.Y)
+            {
+                Configuration.ShowRegistryAdditionPrompt = false;
+                return;
+            }
+
+            try
+            {
+                RegistryKey open = Registry.ClassesRoot.CreateSubKey("*\\shell\\Open with TML.Patcher");
+                RegistryKey command = open.CreateSubKey("command");
+                open.SetValue(null, "Open with TML.Patcher");
+                open.SetValue("icon", Path.Combine(ExePath, "TML.Patcher.Frontend.exe"));
+                command.SetValue(null, $"{Path.Combine(ExePath, "TML.Patcher.Frontend.exe")} \"%1\"");
+            }
+            catch (UnauthorizedAccessException)
+            {
+                Console.WriteLine("Insufficient permissions to modify the registry. Please re-launch with admin permissions to add to your context menu, otherwise press any key to continue.");
+                Console.ReadKey();
+            }
+
+            Configuration.ShowRegistryAdditionPrompt = false;
+
         }
 
         internal static void PreLoadAssemblies()